pyproject.toml
[project]
name = "choptimize"
version = "0.1.0"
description = "A CLI tool to analyze and optimize coding prompts"
readme = "README.md"
authors = [{ name = "darragh0" }]
requires-python = ">=3.13"
dependencies = [
  "google-genai>=0.6.0",
  "python-dotenv>=1.0.0",
  "rich>=13.0.0",
  "rich-argparse>=1.7.2",
]

[project.scripts]
choptimize = "choptimize.__main__:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"
